Return error if data buffer not on BLOCK_SIZE boundary
BZ: 212875

There were two issues with the logic.

One is that the check for proper block size boundary
for AES was done using the block count variable and not
the buffer size variable.

The other is that it is unacceptable to simply increment
the number of blocks if there is a partial block. If
there is a partial block at the end of the buffer; ie;
buffer size is not exactly on block size boundary; then
we need to reject the call and return with zero length
(error condition)

Change-Id: If8f61b5201280acf468df30767d9a55ada5f7ad7
Category: device enablement
Domain: AOSP.Optimization-Bouncycastle
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Mark Allyn <mark.a.allyn@intel.com>